cleaning up src
