+++ src/__tests__/collapse.controlled.jest.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {render, cleanup} from 'react-testing-library';
-import {useCollapse} from '../collapse-hooks';
+import {useCollapse} from '../collapsed';
 // add custom jest matchers from jest-dom
 import 'jest-dom/extend-expect';
